+++ arduino/src/ErrorCode.h
@@ -15,7 +15,7 @@ struct ErrorCode
         // But we don't care about actual values
 
         NO_CARD = 0x80,
-        BAD_CHECKSUM,
+        DATA_CORRUPTED,
         NO_SERIAL,
         WRONG_CARD,
         CARD_IS_FULL,
@@ -33,3 +33,34 @@ struct ErrorCode
     {
     }
 };
+
+#ifdef BUILD_TEST
+
+#include <iostream>
+
+inline std::ostream& operator<<(std::ostream &os, ErrorCode e)
+{
+    switch (e) {
+    case ErrorCode::NO_CARD:
+        os << "No card";
+        break;
+    case ErrorCode::DATA_CORRUPTED:
+        os << "Corrupt";
+        break;
+    case ErrorCode::NO_SERIAL:
+        os << "No serial";
+        break;
+    case ErrorCode::WRONG_CARD:
+        os << "Wrong card";
+        break;
+    case ErrorCode::CARD_IS_FULL:
+        os << "Full";
+        break;
+    default:
+        os << "Error(" << static_cast<int>(e) << ")";
+        break;
+    }
+    return os;
+}
+
+#endif //BUILD_TEST